Union-merge of origin/develop into main per owner directive: nothing
un-wired, verify green before landing, exclude only what provably cannot
pass the strict gate.

Conflict resolution (20 files):
- Source/manifest conflicts resolved toward develop (the green integration
  tip): spiking-network, codex-env, ruvector-core/agenticdb.rs, micro-hnsw-wasm,
  ruvix cap/proof feature decls. Removed stale single-file modules
  (spiking-network learning.rs/router.rs) superseded by develop's dir modules.
- Root Cargo.toml resolved as a strict UNION of members (merged ⊇ main ∪ develop;
  nothing develop or main had as a member was dropped) with members ∩ excludes = ∅.
- Carried develop's audit-ignore (RUSTSEC-2026-0041, zenoh-transitive lz4_flex).

Greened newly-unified crates under `clippy --workspace --all-targets --no-deps -D warnings`:
- ruvector-maxsim, ruvector-agent-memory, ruvector-hnsw-repair: reused the
  verified greening from the (now-superseded) fix/robotics-core-deps line.
- exo-ai-2025 cluster (9 crates) + rvf-examples: ~150 real idiomatic clippy
  fixes (entry/or_default, iterator+enumerate, div_ceil, Range::contains,
  clamp, saturating_sub, …). Only domain-justified #[allow]s (FFI/napi casts).

Excluded — provably cannot pass the host strict gate, source preserved on disk,
develop-aligned (no capability loss):
- RVM bare-metal Pi-5 kernel cluster (ruvix bcm2711/cli/dma/dtb/fs/net/physmem/
  shell/smp): nested no_std kernel; ~262 domain-inappropriate pedantic lints.
- vendor/cognitum-one: vendored upstream SDK whose tests import a renamed crate;
  we do not patch vendored code.

Verification: cargo clippy --workspace --all-targets --no-deps -- -D warnings → 0;
cargo fmt --all --check → clean. 215 workspace members.

+++ examples/rvf/examples/solver_witness.rs
@@ -124,16 +124,12 @@ fn main() {
 
         let phase = if iter < 3 {
             "warmup"
-        } else if residual > convergence_tol as f64 {
+        } else if residual > convergence_tol {
             "converging"
         } else {
             "converged"
         };
-        let converged: u64 = if residual <= convergence_tol as f64 {
-            1
-        } else {
-            0
-        };
+        let converged: u64 = if residual <= convergence_tol { 1 } else { 0 };
 
         // Store residual as fixed-point u64 (multiply by 1e9)
         let residual_fixed = (residual * 1e9) as u64;
@@ -213,7 +209,7 @@ fn main() {
 
             let wtype = if i == 0 {
                 0x01 // PROVENANCE: initial state
-            } else if residuals[i] <= convergence_tol as f64 {
+            } else if residuals[i] <= convergence_tol {
                 0x03 // custom: CONVERGENCE witness
             } else {
                 0x02 // COMPUTATION: intermediate iteration
@@ -398,10 +394,7 @@ fn main() {
     println!("  Deterministic replay:    VERIFIED");
     println!("  Tamper detection:        WORKING");
 
-    let converged_count = residuals
-        .iter()
-        .filter(|&&r| r <= convergence_tol as f64)
-        .count();
+    let converged_count = residuals.iter().filter(|&&r| r <= convergence_tol).count();
     println!(
         "  Converged iterations:    {} / {}",
         converged_count, num_iterations
